+++ CHANGELOG.md
@@ -2,11 +2,14 @@
 ---
 ## [Known Bug]
 - In cases where two annotations overlap on the same slice, updating the nodule table results in the creation of two separate nodules. The expected behavior is to merge the two nodules in such scenarios.
-## [Unreleased]
+## [Unreleased] 
+### Changed
+- Hide the patient which is not inferenced by the model.
 ### Fixed
-- Change loaded history from list to set to speed up the process of loading history
+- Speed up the loading of the nodule table
 - Set the time zone for displaying the last update time on the table to UTC+8 for consistency and accuracy in representation
-
+- Set the time zone for saving log files to UTC+8 for consistency and accuracy in representation
+- After loading the log, it cannot press "Update" button to update the nodule table.
 ## [2.0.9] - 2023-10-13
 ### Added
 - Added User Option: